move githubVerifier to its own file
